requirements.txt
streamlit==1.31.1
tensorflow==2.18.0
numpy>=1.24.3
pandas>=2.1.4
scikit-image>=0.22.0
Pillow>=10.2.0
matplotlib>=3.7.0 
